split BicepNetBaseCommand into two files, logger and pscmdlet
